Moved common functions to setup_commons.py
